refactor!: Extracted utility classes to `lettucemod-utils` module
